Color Game
In this game, the player has to enter the color of the word that appears on the screen and hence the score increases by one, the total time to play this game is 30 seconds. Colors used in this game are Red, Blue, Green, Pink, Black, Yellow, Orange, White, Purple and Brown. The interface will display the names of different colors in different colors. The player has to identify the color and enter the correct color name to win the game. The technology used is Python and Python tkinter GUI library.